+++ MLEM/Formatting/Codes/ShadowCode.cs
@@ -24,10 +24,5 @@ namespace MLEM.Formatting.Codes {
             return false;
         }
 
-        /// <inheritdoc />
-        public override bool EndsHere(Code other) {
-            return other is ShadowCode || other is ResetFormattingCode;
-        }
-
     }
 }